library_management_system/requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
Werkzeug==2.0.1 